requirements.txt

celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.0.6
django-crispy-forms==2.2
djangorestframework==3.15.2
yfinance==0.2.40
